:fire:remove unused file :truck: moved scripts
